successfully connected database and mongoose
