+++ src/backend/db.js
@@ -1,0 +1,12 @@
+// import mongoose from "mongoose"
+// import dotenv from "dotenv"
+
+// dotenv.config()
+
+// try {
+//     const m = await mongoose.connect(process.env.DB_URI)
+//     console.log(m.connection.readyState == 1 ? 'Mongoose Connected' : 'Mongoose failed to connect')
+// }
+// catch (err) {
+//     console.error(err)
+// }
