+++ src/backend/models/bookings.js
@@ -1,18 +1,18 @@
-import mongoose from "mongoose"
-import { User } from "users.js"
-import { Equipment } from "equipment.js"
-import { HireOptions } from "hireOptions.js"
+// import mongoose from "mongoose"
+// import { User } from "users.js"
+// import { Equipment } from "equipment.js"
+// import { HireOption } from "hireOptions.js"
 
 
-const bookingSchema = new mongoose.Schema({
-    user: { type: mongoose.Schema.Types.ObjectId, ref: User },
-    equipment: { type: mongoose.Schema.Types.ObjectId, ref: Equipment },
-    date: { type: Date, required: true },
-    time: { type: Date, required: true },
-    hireOptions: { type: mongoose.Schema.Types.ObjectId, ref: HireOptions },
-    quantity: { type: Number, required: true }
-})
+// const bookingSchema = new mongoose.Schema({
+//     user: { type: mongoose.Schema.Types.ObjectId, ref: User },
+//     equipment: { type: mongoose.Schema.Types.ObjectId, ref: Equipment },
+//     date: { type: Date, required: true },
+//     time: { type: Date, required: true },
+//     hireOptions: { type: mongoose.Schema.Types.ObjectId, ref: HireOption },
+//     quantity: { type: Number, required: true }
+// })
 
-const Booking = mongoose.model('Booking', bookingSchema);
+// const Booking = mongoose.model('Booking', bookingSchema);
 
-export default Booking
+// export default Booking
