+++ src/backend/models/equipment.js
@@ -1,19 +1,19 @@
-import mongoose from "mongoose"
-import HireOption from "hireOptions.js"
+// import mongoose from "mongoose"
+// import { HireOption } from "hireOptions.js"
 
-const equipmentSchema = new mongoose.Schema({
-    item: { type: String, required: true },
-    quantity: { type: Number, required: true },
-    hireOption: [
-        {
-            hireOption: {type: mongoose.Schema.Types.ObjectId, ref: HireOption, required: true},
-            price: { type: Number, required: true }
-        }
-    ],
-    image: { type: String, required: true},
-})
+// const equipmentSchema = new mongoose.Schema({
+//     item: { type: String, required: true },
+//     quantity: { type: Number, required: true },
+//     hireOption: [
+//         {
+//             hireOption: {type: mongoose.Schema.Types.ObjectId, ref: HireOption, required: true},
+//             price: { type: Number, required: true }
+//         }
+//     ],
+//     image: { type: String, required: true},
+// })
 
 
-const Equipment = mongoose.model("Equipment", equipmentSchema)
+// const Equipment = mongoose.model("Equipment", equipmentSchema)
 
-export default Equipment
+// export default Equipment
