+++ src/backend/models/users.js
@@ -1,14 +1,14 @@
-import mongoose from "mongoose"
+// import mongoose from "mongoose"
 
 
-const userSchema = new mongoose.Schema({
-    firstName: { type: String, required: true },
-    lastName: { type: String, required: true },
-    email: { type: String, required: true, unique: true },
-    password: { type: String, required: true },
-    isAdmin: { type: Boolean, required: true, default: false}
-})
+// const userSchema = new mongoose.Schema({
+//     firstName: { type: String, required: true },
+//     lastName: { type: String, required: true },
+//     email: { type: String, required: true, unique: true },
+//     password: { type: String, required: true },
+//     isAdmin: { type: Boolean, required: true, default: false}
+// })
 
-const User = mongoose.model("User", userSchema)
+// const User = mongoose.model("User", userSchema)
 
-export default User
+// export default User
